10.09.2022

+++ app/src/main/java/uz/jbnuu/tsc/model/attendance/AttendanceData.kt
@@ -14,5 +14,5 @@ data class AttendanceData(
     val employee: Employee?,
     val absent_on: Int?,
     val absent_off: Int?,
-    val lesson_date: Int?
+    val lesson_date: Long?
 )
