+++ app/src/main/java/uz/jbnuu/tsc/model/schedule/ScheduleData.kt
@@ -3,9 +3,9 @@ package uz.jbnuu.tsc.model.schedule
 data class ScheduleData(
     val subject: Subject?,
     val _week: Int?,
-    val weekStartTime: Int?,
-    val weekEndTime: Int?,
-    val lesson_date: Int?,
+    val weekStartTime: Long?,
+    val weekEndTime: Long?,
+    val lesson_date: Long?,
     val additional: String?,
     val auditorium: Auditorium?,
     val trainingType: Auditorium?,
